Cleaned up and refactored all the model code to rcr::detection_model
- Updated the model binary
- Removed code that was previously duplicated in rcr-train and rcr-detect (the solver and model class)
- Cleaned up includes in rcr-detect
- Removed using's in model.hpp
